# Conflicts:
#	apps/app-Demo-SLProject/iOS/CMakeLists.txt
#	apps/app-Demo-WAI/GLFW/AppDemoMainGLFW.cpp
#	apps/app-Demo-WAI/iOS/CMakeLists.txt
#	apps/app-Demo-WAI/iOS/ErlebARViewController.mm
#	apps/app-Demo-WAI/source/ErlebAR.cpp
#	apps/app-Demo-WAI/source/scenes/AppWAIScene.cpp
#	apps/app-Demo-WAI/source/views/AreaTrackingView.cpp
#	apps/app-Demo-WAI/source/views/AreaTrackingView.h
#	apps/source/sens/SENSARCore.h
#	apps/source/sens/android/SENSNdkARCore.cpp
#	apps/source/sens/android/SENSNdkARCore.h
#	cmake/CompileOptions.cmake
#	lib-SLProject/source/CV/CVFeatureManager.cpp

+++ lib-WAI/source/OrbSlam/LocalMapping.cpp
@@ -639,7 +639,7 @@ void LocalMapping::SearchInNeighbors(LocalMap& lmap)
                 continue;
             if (pMP->isBad() || pMP->mnMarker[FUSE_CANDIDATE_KF] == lmap.refKF->mnId)
                 continue;
-            pMP->mnMarker[FUSE_CANDIDATE_KF] = lmap.refKF->mnId;
+            pMP->mnMarker[FUSE_CANDIDATE_KF] = (int)lmap.refKF->mnId;
             vpFuseCandidates.push_back(pMP);
         }
     }
@@ -721,7 +721,7 @@ void LocalMapping::SearchInNeighbors(WAIKeyFrame* frame)
                 continue;
             if (pMP->isBad() || pMP->mnMarker[FUSE_CANDIDATE_KF] == frame->mnId)
                 continue;
-            pMP->mnMarker[FUSE_CANDIDATE_KF] = frame->mnId;
+            pMP->mnMarker[FUSE_CANDIDATE_KF] = (int)frame->mnId;
             vpFuseCandidates.push_back(pMP);
         }
     }
